# postings: use Loser Tree for merge