@bboreham 2023-11-13
Yes it can be merged.  I had hoped to get the implementation closer between Loki and Prometheus, but currently there is too much of a performance loss to do that.

Also you might not want to take a dependency on something from my personal account.  There are precedents - alecthomas, cespare,  dennwc, klauspost.