@yeya24 2024-01-18
Hi @bboreham, may I know if you have any benchmark for mergedPostings `Seek` method? Do we expect to see similar improvements as the provided benchmark