@jub0bs 2024-04-01
For information, Bryan's talk is now available at https://www.youtube.com/watch?v=AmLtlXEo4UU